% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allele.freq}
\alias{allele.freq}
\title{Frequencies of SNPs.}
\format{A data frame with 6 rows and 3 columns(variables):
\describe{
  \item{SNP}{name of SNP}
  \item{major.frequency}{frequency of major allele}
  \item{minor.frequency}{frequency of minor allele}
}}
\usage{
data(allele.freq)
}
\description{
A dataset containing sample allele frequencies of SNPs.
allele.freq.
}
\keyword{datasets}
