% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcuheritability.R
\name{calc.herit}
\alias{calc.herit}
\title{Calculate heritability}
\usage{
calc.herit(gene.coef, freq, noise.var, Dskim = 0, ...)
}
\arguments{
\item{gene.coef}{a vector of 8 genetic parameters.
Each element includes 8 genetic parameters (regression
coefficient)
\eqn{\beta_{Gw_{t}}}, t in (1, 2, \eqn{\cdots}{...}, 8).}

\item{freq}{a dataframe with the allele frequencies.}

\item{noise.var}{variance of noise to generate the random noise.}

\item{Dskim}{the coefficient of linkage disequilibrium. Default is 0 (no LD).}

\item{...}{not used.}
}
\value{
heritability.
}
\description{
Calculate heritability
}
\author{
Beibei Jiang \email{beibei_jiang@psych.mpg.de}
}
