% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readpars.R
\name{check.snp.par}
\alias{check.snp.par}
\title{Check the number of the SNPs involved in epistasis and main effects}
\usage{
check.snp.par(genetic.pars, nphe, ...)
}
\arguments{
\item{genetic.pars}{a data.frame or a matrix containing the parameter
information for the main effect: additive and dominance.}

\item{nphe}{number of phenotypes.}

\item{...}{not used.}
}
\value{
NULL --- will stop if test fails.
}
\description{
Check the number of the SNPs set in the simulation parameters.
}
\details{
The number of SNPs involved in main effects should be the same as the number
of SNPs involved in epistasis.
}
\examples{
check.snp.par(genepars, nphe = 2)
}
\author{
Beibei Jiang \email{beibei_jiang@psych.mpg.de}
}
