% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epistasis.pars}
\alias{epistasis.pars}
\title{Parameter settings of epistatic effects.}
\format{A data frame with 6 rows and 3 columns(variables):
\describe{
  \item{SNPA}{first SNP}
  \item{SNPB}{second SNP}
  \item{additive_additive}{coefficient for additive-additive interaction}
  \item{additive_dominance}{coefficient for additive-dominance interaction}
  \item{dominance_additive}{coefficient for dominance-additive interaction}
  \item{dominance_dominance}{coefficient for dominance-dominance interaction}
}}
\usage{
data(epistasis.pars)
}
\description{
A dataset containing the parameter settings for epistatic effects.
epistasis.pars.
}
\keyword{datasets}
