% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene.coefficients}
\alias{gene.coefficients}
\title{Coefficients of genetic effects.}
\format{A list with 3 elements:
\describe{
\item{epi.par1}{a data frame with 1 row and 10 variables:
                \describe{
                  \item{SNPA}{first SNP}
                  \item{SNPB}{second SNP}
                  \item{additiveA}{coefficient for additive effect at locus A}
                  \item{dominanceA}{coefficient for dominance effect at locus A}
                  \item{additiveB}{coefficient for additive effect at locus B}
                  \item{dominanceB}{coefficient for dominance effect at locus B}
                  \item{additive_additive}{coefficient for additive-additive interaction}
                  \item{additive_dominance}{coefficient for additive-dominance interaction}
                  \item{dominance_additive}{coefficient for dominance-additive interaction}
                  \item{dominance_dominance}{coefficient for dominance-dominance interaction}
                  }}
\item{epi.par2}{a data frame with 1 row and 10 variables:
                \describe{
                  \item{SNPA}{first SNP}
                  \item{SNPB}{second SNP}
                  \item{additiveA}{coefficient for additive effect at locus A}
                  \item{dominanceA}{coefficient for dominance effect at locus A}
                  \item{additiveB}{coefficient for additive effect at locus B}
                  \item{dominanceB}{coefficient for dominance effect at locus B}
                  \item{additive_additive}{coefficient for additive-additive interaction}
                  \item{additive_dominance}{coefficient for additive-dominance interaction}
                  \item{dominance_additive}{coefficient for dominance-additive interaction}
                  \item{dominance_dominance}{coefficient for dominance-dominance interaction}
                  }}
\item{epi.par3}{a data frame with 1 row and 10 variables:
                \describe{
                  \item{SNPA}{first SNP}
                  \item{SNPB}{second SNP}
                  \item{additiveA}{coefficient for additive effect at locus A}
                  \item{dominanceA}{coefficient for dominance effect at locus A}
                  \item{additiveB}{coefficient for additive effect at locus B}
                  \item{dominanceB}{coefficient for dominance effect at locus B}
                  \item{additive_additive}{coefficient for additive-additive interaction}
                  \item{additive_dominance}{coefficient for additive-dominance interaction}
                  \item{dominance_additive}{coefficient for dominance-additive interaction}
                  \item{dominance_dominance}{coefficient for dominance-dominance interaction}
                  }}
}}
\usage{
data(gene.coefficients)
}
\description{
A dataset containing the regression coefficients of genetic effects.
gene.coefficients.
}
\keyword{datasets}
