% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genepars}
\alias{genepars}
\title{Parameter settings for simulation.}
\format{A list with 7 elements:
\describe{
\item{P1mean}{A data frame with 1 row and 1 variable:
              \describe{
                \item{mean}{\eqn{\beta_{0}}: coefficient parameter of "basic" genetic effects in \eqn{G_{ij}=\beta_{0}+\sum_{t=1}^{8}\beta_{G_{w_{t}}}w_{tij}}. }
                }}
\item{P1main}{A data frame with 6 rows and 3 variables:
              \describe{
                \item{SNP}{SNP name}
                \item{additive}{coefficient of additive effect}
                \item{dominance}{coefficient of dominance effect}
                }}
\item{P1epistasis}{A data frame with 3 rows and 6 variables:
                  \describe{
                    \item{SNPA}{first SNP}
                    \item{SNPB}{second SNP}
                    \item{additive_additive}{coefficient for additive-additive interaction}
                    \item{additive_dominance}{coefficient for additive-dominance interaction}
                    \item{dominance_additive}{coefficient for dominance-additive interaction}
                    \item{dominance_dominance}{coefficient for dominance-dominance interaction}
                    }}
\item{P1heritability}{A data frame with 1 row and 1 variable:
                     \describe{
                        \item{heritability}{expected heritability}
                        }}
\item{P2mean}{A data frame with 1 row and 1 variable:
              \describe{
                \item{mean}{mean of genetic effect}
                }}
\item{P2main}{A data frame with 6 rows and 3 variables:
              \describe{
                \item{SNP}{SNP name}
                \item{additive}{coefficient of additive effect}
                \item{dominance}{coefficient of dominance effect}
                }}
\item{P2epistasis}{A data frame with 3 rows and 6 variables:
                  \describe{
                    \item{SNPA}{first SNP}
                    \item{SNPB}{second SNP}
                    \item{additive_additive}{coefficient for additive-additive interaction}
                    \item{additive_dominance}{coefficient for additive-dominance interaction}
                    \item{dominance_additive}{coefficient for dominance-additive interaction}
                    \item{dominance_dominance}{coefficient for dominance-dominance interaction}
                    }}
}}
\usage{
data(genepars)
}
\description{
A dataset containing the parameters for a simulation.
genepars.
}
\keyword{datasets}
