% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filewrite.R
\name{list2frame}
\alias{list2frame}
\title{Convert list to data.frame}
\usage{
list2frame(x, ...)
}
\arguments{
\item{x}{a list. In this package, it is used as a list includes simulated
phenotypes. One element per block.
Each element is a dataframe with SNP names and individuals.}

\item{...}{not used.}
}
\value{
a data.frame.
}
\description{
Convert list to data.frame.
}
\examples{
x <- list(test1=matrix(rnorm(1000), ncol=2), test2=matrix(rnorm(1000), ncol=2))
str(x)
x.new <- list2frame(x)
str(x)
}
\author{
Beibei Jiang \email{beibei_jiang@psych.mpg.de}
}
