% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maineff.pars}
\alias{maineff.pars}
\title{Parameter settings of main effects (additive effect and dominance).}
\format{A data frame with 6 rows and 3 columns(variables):
\describe{
  \item{SNP}{name of SNP}
  \item{additive}{coefficient of additive effect}
  \item{dominance}{coefficient of dominance effect}
}}
\usage{
data(maineff.pars)
}
\description{
A dataset containing the parameter settings for main effects.
maineff.pars.
}
\keyword{datasets}
