% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filewrite.R
\name{pars.writer}
\alias{pars.writer}
\title{Write current parameters to file}
\usage{
pars.writer(genetic.pars, fname = "usedpars.txt", ...)
}
\arguments{
\item{genetic.pars}{list of simulation parameters. One element per block of
\code{file}. Each element is a dataframe with SNP names
and model parameters.}

\item{fname}{filename of the setting for simulation (for recording).}

\item{...}{not used.}
}
\description{
Write out the information about the parameters for simulation.
}
\examples{
pars.writer(genepars)
}
\author{
Beibei Jiang \email{beibei_jiang@psych.mpg.de}
}
