% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filewrite.R
\name{phe.writer}
\alias{phe.writer}
\title{Write phenotypes}
\usage{
phe.writer(phe, onefile = TRUE, fname = "simu.pheno", ...)
}
\arguments{
\item{phe}{a data.frame or a matrix of simulated phenotypes.
Each column is a phenotype.}

\item{onefile}{whether to create just one file for all phenotypes (default) or
one per phenotype}

\item{fname}{filename of the phnotype(s).}

\item{...}{not used.}
}
\description{
Write out the simulated phenotypes.
}
\examples{
phe <- matrix(rnorm(1000), ncol = 2)
colnames(phe) <- c("p1", "p2")
phe.writer(phe)
}
\author{
Beibei Jiang \email{beibei_jiang@psych.mpg.de} and
        Benno Pütz \email{puetz@psych.mpg.de}
}
